Add semantic versioning for endpoints
Sometimes it's important for the client to know if they are
complaint with the changes in the endpoint (or even to get
know that some changes even occurred).

We can help client to decide by introducing Endpoint Semantic
Versioning according to the Semantic Versioning v. 2.0.0
(http://semver.org/).

Versions of available endpoints can be retrieved by /_version
or found in the X-Endpoint-Version response header.